Added functionality to the trainer
Unfortunately, using pyshark to sniff the traffic has turned out to be a bad idea. It crashes frequently when running multiple sniffing sessions, and tshark continues to run in the background even when the program is not sniffing. Have been unable to run a long training session. Will likely need to swap it out for Scapy instead