Rewrote the TLS parser and got everything working again
Lots more features I'd like to add, and I'm pretty sure I'm incorrectly parsing some of the application data packets still (will often receive several of them in a session that are too small to contain an IV or MAC for the crypto. Still it when using it in sniffer mode, it has become much more reliable.